% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasegawa2016.R
\name{cor.0}
\alias{cor.0}
\alias{cor.1}
\title{Predicted cross-test correlation}
\usage{
cor.0(rho1, gamma1, rho2, gamma2, lambda, R, p, t.star)

cor.1(rho1, gamma1, rho2, gamma2, lambda, theta, eps, R, p, t.star)
}
\arguments{
\item{rho1}{First power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}

\item{gamma1}{Second power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}

\item{rho2}{First power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}

\item{gamma2}{Second power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}

\item{lambda}{Event hazard for the control arm.}

\item{R}{End of the accrual period.}

\item{p}{Treatment assignment probability.}

\item{t.star}{Time point we pause the study to check the cumulative information under the null.}

\item{theta}{Hazard ratio after the change point (before the change point HR should be 1).}

\item{eps}{Change point.}
}
\description{
These two functions are to predict the correlation between two weighted log-rank tests at certain time \code{t.star} under either the null hypothesis (using \code{cor.0}) or the alternative hypothesis (using \code{cor.1}).
}
\details{
These two functions are designed to calculate the predicted correlation between the two weighted log-rank tests at time \code{t.star} under the two hypotheses. The null hypothesis is an exponential distribution for both the treatment and control arms with hazard \code{lambda}, while the alternative hypothesis has the control group following an exponential distribution with hazard \code{lambda}, and the treatment group following a piece-wise exponential distribution with hazard \code{lambda} before \code{eps}, but a hazard \code{theta} times \code{lambda} after \code{eps}.
}
\references{
Hasegawa, T. (2016). Group sequential monitoring based on the weighted log‐rank test statistic with the Fleming–Harrington class of weights in cancer vaccine studies. Pharmaceutical statistics, 15(5), 412-419.
}
\author{
Lili Wang.
}
