% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasegawa2016.R
\name{FH.frac.cal}
\alias{FH.frac.cal}
\title{Information fraction for Fleming-Harrington weighted log-rank test}
\usage{
FH.frac.cal(data, t_vec, I_max, rho, gamma, trimmed)
}
\arguments{
\item{data}{There are two possible structures allowed for this input data. The first type needs to have \code{trimmed=F} and include variables: a \code{treatment} variable with "experimental" denoting treatment group, \code{cnsr} variable with value 1 denoting censoring, \code{ct} variable denoting event time from the origin of the study, which equals the sum of entering time \code{enterT} and the survival time (time to event or censoring). A dataset simulated from from R package \href{https://github.com/keaven/nphsim}{nphsim} should fit the first type well enough (see the example1). The second type can be any data.frame or data.table output from a \code{data.trim} function, including variables: \code{ct} denoting event time from the origin of the study or the sum of entering time and the survival time, \code{survival} denoting the survival time or time to event/censoring, \code{delta} as an event indicator, \code{enterT} as entering time (example 2). For the second type, we set \code{trimmed=T} to avoid extra computations, but should be fine if \code{trimmed=F}.}

\item{t_vec}{Follow-up time since the origin of the study (not that it's not following the survival time scale, but following the calendar time scale ), which could be a vector, to measure the information fraction for these time points.}

\item{I_max}{The evaluated \eqn{I_{max}}, which returned from function \code{I.1} or \code{I.0} by setting the \code{t.start=tau}, where \code{tau} is the end of the study.}

\item{rho}{First power parameter for the Fleming-Harrington weight which weighs on the early departures: \eqn{S(t^-)^\rho(1-S(t^-))^\gamma}.}

\item{gamma}{Second power parameter for the Fleming-Harrington weight which weighs on the late departures: \eqn{S(t^-)^\rho(1-S(t^-))^\gamma}.}

\item{trimmed}{Logical indicator to show whether the \code{data} input has been "trimmed" by \code{data.trim} and \code{data.trim.d} before: adding variables like \code{delta} indicating events (=1), and \code{trt} distringuishing the treatment group (=1) from the control group (=0)}
}
\value{
This function returns a vector of information fractions corresponding to the input time vector \code{t_vec}.
}
\description{
Monitor the raction for Fleming-Harrington weighted log-rank test for a vector of time points
}
\details{
Calculation the information fraction for Fleming-Harrington family weighted log-rank tests using the monitored estimated information for numerator, and the predicted information \eqn{I_{max}} as denominator.
}
\examples{
# install.packages("devtools")
# library(devtools)
# install_github("keaven/nphsim")
library(nphsim)
eps<-2 # delayed effect
p<-0.5 #treatment assignment
b<-30 # an intrinsic parameter to decide the number of intervals per time unit
tau<- 18 # end of the study
R<-14 # accrual period [0,R]
omega<- tau-R
lambda<-log(2)/6 # control group risk hazard
theta<-0.7 # hazard ratio
lambda.trt<- lambda*theta #hazard after the change point for the treatment arm
rho<- 0 # parameter for the weights
gamma<-1 #parameter for the weights
alpha<-0.025 #type 1 error
beta<-0.1 #type 2 error
# First we decide the sample size:
size_FH <- sample.size_FH(eps,p,b,tau,omega,lambda,lambda.trt,rho, gamma,alpha,beta)
n_FH <-size_FH$n
n_event_FH<-size_FH$n_event
accrual.rt<-n_FH/R # the needed arrual rate
#Generate data accordingly, use eta=1e-5 to inhibit censoring
data_temp <- nphsim(nsim=1,lambdaC=lambda, lambdaE = c(lambda,lambda.trt),
                  ssC=ceiling(n_FH*(1-p)),intervals = c(eps),ssE=ceiling(n_FH*p),
                                   gamma=accrual.rt, R=R, eta=1e-5, fixEnrollTime = TRUE)$simd

# Example 1 for FH.frac.cal: Set trimmed=F and work on the crude dataset from nphsim()
inf_frac_vec1<-FH.frac.cal(data_temp,c(10,15,18),I_denom,rho,gamma,trimmed=F)
inf_frac_vec1

# Example 2 for FH.frac.cal: First trim the data before inputting into FH.frac.cal() setting trimmed=T, and obtain the whole spectrum.
I_denom<-I.1(rho, gamma,lambda,theta,eps,R,p,t.star=tau)*n_FH
tau.star=21 #in case the ratio=1 when t>tau
#Trim the data
data_temp2 <-data.trim(tau.star,data_temp)
t_seq <- seq(0.1,tau_star,0.1) # the time series to check the information fraction
inf_frac_vec2<-FH.frac.cal(data_temp2,t_seq,I_denom,rho,gamma,trimmed=T)
# WLRT at the interim
interim_index<- which.min(abs(inf_frac_vec2-0.6))
interim_time<-t_seq[interim_index]
interim_frac<-inf_frac_vec2[interim_index]
 # WLRT at the final
 final_index<- which.min(abs(inf_frac_vec2-1))
 final_time<-t_seq[final_index]
 final_frac<-inf_frac_vec2[final_index]
}
\references{
Hasegawa, T. (2016). Group sequential monitoring based on the weighted log‐rank test statistic with the Fleming–Harrington class of weights in cancer vaccine studies. Pharmaceutical statistics, 15(5), 412-419.
}
\seealso{
\code{\link{data.trim}}
}
\author{
Lili Wang
}
