% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLRT.R
\name{FH.table}
\alias{FH.table}
\title{Basic function for Fleming-Harrington family weighted log-rank tests}
\usage{
FH.table(survival, delta, trt, rho, gamma)
}
\arguments{
\item{survival}{Time to event or censoring.}

\item{delta}{Event indicator: 1 for observed cases, 0 for censored cases.}

\item{trt}{Treatment assignment indicator: 1 for treatment group, 0 for control group.}

\item{rho, gamma}{Parameters for Fleming-Harrington family with \eqn{W(t)=S^\rho(t^-)(1-S(t^-))^\gamma}.}
}
\value{
Build a table for Fleming-Harrington log-rank test which ouputs \emph{ordered} \code{survival} as follow-up times, \code{Surv} as predictable survival functions \eqn{S(t^-)}, \code{Surv.exact} as exact survival functions \eqn{S(t)}, \code{delta} as event indicators, \code{trt} as treatment assignement (treated=1, control=0), \code{weight} as weight functon calcualted from the predictable survival functions \code{Surv}.

In addition, the output also include \code{O1} as the observed events from the treatment arm, \code{E1} as the expected events from the treatment arm, \code{Cov} as the estimated variance without considering the weights.
}
\description{
Basic function to build the table for the calculation of the Fleming-Harrington family of weighted log-rank tests.
}
\author{
Lili Wang
}
