% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLRT.R
\name{FH.test}
\alias{FH.test}
\title{Fleming-Harrington weighted log-rank tests}
\usage{
FH.test(survival, delta, trt, rho, gamma)
}
\arguments{
\item{survival}{Time to event or censoring.}

\item{delta}{Event indicators.}

\item{trt}{Treatment assignment indicator with 1 denoting the treated group, and 0 denoting the placebo group.}

\item{rho}{First power parameter for the Fleming-Harrington weight which weighs on the early departures: \eqn{S(t^-)^\rho(1-S(t^-))^\gamma}.}

\item{gamma}{Second power parameter for the Fleming-Harrington weight which weighs on the late departures: \eqn{S(t^-)^\rho(1-S(t^-))^\gamma}.}
}
\value{
A list 3 different components
  \item{O1}{Observed number of weighted events (with a multiplication of corresponding weights) in the treatment arm.}
  \item{E1 }{Expected number of weighted events (with a multiplication of corresponding weights) in the treatment arm.}
  \item{Z}{Weighted log-rank test statistic.}
}
\description{
Calculating the Fleming-Harrington weighted log-rank tests
}
\examples{
# Examples for FH.test and WLR.test
set.seed(12345)
data_temp<- nphsim(nsim=1,lambdaC=log(2)/6, lambdaE = c(log(2)/6,log(2)/6*0.7), ssC=250, intervals = c(2),ssE=250, gamma=500/14, R=14, eta=1e-5, fixEnrollTime = TRUE)$simd
data_final<-data.trim.d(100,data_temp)[[1]]
rho=1
gamma=0
# compare the 3 different ways below:
#library(survival)
sqrt(survdiff(Surv(survival,delta)~trt, data =data_final,rho=rho)$chisq)
FH.test(survival=data_final$survival,delta=data_final$delta,trt=data_final$trt,rho=rho,gamma=gamma)
WLR.test(survival=data_final$survival,delta=data_final$delta,trt=data_final$trt,w=function(...){survKM_minus(...)^rho*(1-survKM_minus(...))^gamma})

}
\seealso{
\code{\link{WLR.test}}
}
\author{
Lili Wang
}
