% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasegawa2016.R
\name{getc}
\alias{getc}
\alias{uv}
\alias{v}
\alias{u}
\alias{h1}
\alias{h0}
\alias{h.tilde}
\title{Basic functions}
\usage{
getc(theta, lambda, eps)

uv(e, k, lambda, R, t.star)

v(e, k, lambda)

u(e, k, lambda, R, t.star)

h1(k1, k2, lambda, theta, eps, R, t.star)

h0(k1, k2, lambda, theta, eps, R, t.star)

h.tilde(m, lambda, theta, eps, R, p, t.star)
}
\arguments{
\item{theta}{Hazard ratio after the change point (before the change point HR should be 1).}

\item{lambda}{Event hazard for the control arm.}

\item{eps}{Change point.}

\item{e}{Some convenience parameter to control the change point, which is usually set to be eps or tau}

\item{k, k1, k2, m}{Parameters to control the exponential power of the survival functions (the control arm for the null hypothesis or the weighted sum of two arms for the alternative hypotheiss).}

\item{R}{End of the accrual period.}

\item{t.star}{Time point we pause the study to check the cumulative results.}

\item{p}{Treatment assignment probability.}
}
\value{
\code{getc} returns the \eqn{\exp(-\lambda*\epsilon*(1-\theta))} which is a multiplier for the survival and hazard of the treatment arm after the change point \code{eps}.
}
\description{
Some basic functions for information prediction.
}
\details{
To prepare the values for the prediction of information values. The control arm is following an exponential with rate \code{lambda}, the treatment arm is piece-wise exponential with hazard ratio with respect to the control arm to be 1 before the changing point \code{eps}, and \code{theta} after the change point.
}
\author{
Lili Wang
}
