% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasegawa2016.R
\name{I.0}
\alias{I.0}
\alias{I.0.cov}
\title{Precdict information/covariance under null hypothesis}
\usage{
I.0(rho, gamma, lambda, R, p, t.star)

I.0.cov(rho1, gamma1, rho2, gamma2, lambda, R, p, t.star)
}
\arguments{
\item{rho}{First power parameter for the Fleming-Harrington weight which weighs on the early departures: \eqn{S(t^-)^\rho(1-S(t^-))^\gamma}.}

\item{gamma}{Second power parameter for the Fleming-Harrington weight which weighs on the late departures: \eqn{S(t^-)^\rho(1-S(t^-))^\gamma}.}

\item{lambda}{Event hazard for the control arm.}

\item{R}{End of the accrual period.}

\item{p}{Treatment assignment probability.}

\item{t.star}{Time point we pause the study to check the cumulative information under the null.}

\item{rho1, rho2}{First power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}

\item{gamma1, gamma2}{Second power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}
}
\description{
Calulcation of the information/covariance based on a presumed survival function under the null.
}
\details{
This function is prepared to calculate the predicted information/covariance purely based on the assumed survival function under the null hypothesis: an exponential distribution with hazard \code{lambda}.
}
\references{
Hasegawa, T. (2016). Group sequential monitoring based on the weighted log‐rank test statistic with the Fleming–Harrington class of weights in cancer vaccine studies. Pharmaceutical statistics, 15(5), 412-419.
}
\seealso{
\code{\link{I.1}}
}
\author{
Lili Wang.
}
