% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasegawa2016.R
\name{I.1}
\alias{I.1}
\alias{I.1.cov}
\title{Predicted information/covariance under the alternative hypothesis}
\usage{
I.1(rho, gamma, lambda, theta, eps, R, p, t.star)

I.1.cov(rho1, gamma1, rho2, gamma2, lambda, theta, eps, R, p, t.star)
}
\arguments{
\item{rho}{First power parameter for the Fleming-Harrington weight which weighs on the early departures: \eqn{S(t^-)^\rho(1-S(t^-))^\gamma}.}

\item{gamma}{Second power parameter for the Fleming-Harrington weight which weighs on the late departures: \eqn{S(t^-)^\rho(1-S(t^-))^\gamma}.}

\item{lambda}{Event hazard for the control arm.}

\item{theta}{Hazard ratio after the change point (before the change point HR should be 1).}

\item{eps}{Change point.}

\item{R}{End of the accrual period.}

\item{p}{Treatment assignment probability.}

\item{t.star}{Time point we pause the study to check the cumulative information under the null.}

\item{rho1}{First power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}

\item{gamma1}{Second power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}

\item{rho2}{First power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}

\item{gamma2}{Second power parameters for the two Fleming-Harrington weights, defined for covariance calculation.}
}
\description{
Calulcation of the information/covariance based on a presumed survival function under the alternative hypothesis.
}
\details{
This function is prepared to calculate the predicted information/covariance purely based on the assumed survival function under the alternaitve hypothesis: the control group is following an exponential distribution with hazard \code{lambda}, while the treatment group is following a piece-wise exponential distribution with same hazard before \code{eps}, but a hazard equals \code{theta} times the \code{lambda} after \code{eps}.
}
\references{
Hasegawa, T. (2016). Group sequential monitoring based on the weighted log‐rank test statistic with the Fleming–Harrington class of weights in cancer vaccine studies. Pharmaceutical statistics, 15(5), 412-419.
}
\seealso{
\code{\link{I.0}}
}
\author{
Lili Wang.
}
