% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLRT.R
\name{logrank.table}
\alias{logrank.table}
\title{Basic function for standard log-rank test}
\usage{
logrank.table(survival, delta, trt)
}
\arguments{
\item{survival}{Time to event or censoring.}

\item{delta}{Event indicator: 1 for observed cases, 0 for censored cases.}

\item{trt}{Treatment assignment indicator: 1 for treatment group, 0 for control group.}
}
\value{
Build a table for log-rank test which ouputs \emph{ordered} \code{survival} as follow-up times, \code{delta} as event indicators,\code{trt} as treatment assignement (treated=1, control=0), \code{Y} as the at-risk numbers, \code{P1} as the proportion of treated set, \code{P0} as the proportion of the control set.

In addition, the output also include \code{O1} as the observed events from the treatment arm, \code{E1} as the expected events from the treatment arm, \code{Cov} as the estimated variance.
}
\description{
Build the table for log-rank test calculation.
}
\seealso{
\code{\link{FH.test}}, \code{\link{I_t.2}}, \code{\link{WLR.test.cov}},\code{\link{WLR.test.cor}}
}
\author{
Lili Wang
}
