% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLRT.R
\name{survKM_minus}
\alias{survKM_minus}
\alias{survKM_exact}
\title{Calculate the survival functions}
\usage{
survKM_minus(v, survival, delta)

survKM_exact(v, survival, delta)
}
\arguments{
\item{v}{Time vector to give the corresponding survival functions.}

\item{survival}{Input follow-up times.}

\item{delta}{Input event indicators.}
}
\value{
\code{survKM_minus} returns the predictable one \eqn{S(t^-)}, and \code{survKM_exact} returns \eqn{S(t)}.
}
\description{
Calculate the survival function, either the predictable one \eqn{S(t^-)} using \code{survKM_minus} or \eqn{S(t)} using \code{survKM_exact}.
}
\author{
Lili Wang
}
