% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WLRT.R
\name{WLR.test}
\alias{WLR.test}
\title{Weighted log-rank tests with any input weight}
\usage{
WLR.test(survival, delta, trt, w = function(v, ...) {     1 })
}
\arguments{
\item{survival}{Time to event or censoring.}

\item{delta}{Event indicator: 1 for observed cases, 0 for censored cases.}

\item{trt}{Treatment assignment indicator: 1 for treatment group, 0 for control}

\item{w}{Weight function, with default to be 1, which is similar to the use of input arbitray weight in \code{\link{WLR.test.cov}}. Please also refer to the examples as well.}
}
\description{
Weighted log-rank test for any input weight function.
}
\examples{
# Examples for FH.test and WLR.test
set.seed(12345)
data_temp<- nphsim(nsim=1,lambdaC=log(2)/6, lambdaE = c(log(2)/6,log(2)/6*0.7), ssC=250, intervals = c(2),ssE=250, gamma=500/14, R=14, eta=1e-5, fixEnrollTime = TRUE)$simd
data_final<-data.trim.d(100,data_temp)[[1]]
rho=1
gamma=0
# compare the 3 different ways below:
#library(survival)
sqrt(survdiff(Surv(survival,delta)~trt, data =data_final,rho=rho)$chisq)
FH.test(survival=data_final$survival,delta=data_final$delta,trt=data_final$trt,rho=rho,gamma=gamma)
WLR.test(survival=data_final$survival,delta=data_final$delta,trt=data_final$trt,w=function(...){survKM_minus(...)^rho*(1-survKM_minus(...))^gamma})

}
\seealso{
\code{\link{FH.test}}
}
\author{
Lili Wang
}
