% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hasegawa2014.R
\name{avg.haz}
\alias{avg.haz}
\title{Calcualte the average hazard ratios}
\usage{
avg.haz(theta, eps, lambda, p = 1/2)
}
\arguments{
\item{theta}{hazard ratio after eps between the treatment and the control group, assuming that the hazard rato is 1 before eps.}

\item{eps}{The change point, before which, the hazard ratio is 1, and after which, the hazard ratio is theta.}

\item{lambda}{The constant hazard for the control arm.}

\item{p}{Treatment assignment probability.}
}
\description{
Calculate the average hazard ratios according to Kalbfleisch and Prentice (1981) or in the paper Hasegawa (2014) for piece-wise exponential survival functions (only one change point \code{eps}).
}
\examples{
\dontrun{
# test
lambda=log(2)/6
theta=0.7
eps=2
avg.haz(theta,eps,lambda)
}
}
\references{
Hasegawa, T. (2014). Sample size determination for the weighted log‐rank test with the Fleming–Harrington class of weights in cancer vaccine studies. Pharmaceutical statistics, 13(2), 128-135.
}
\author{
Lili Wang
}
